requirements.txt
fastapi
uvicorn
langchain
langchain-groq
groq
openai
python-dotenv
pydantic
pypdf
re
json
datetime